changed DB

+++ reddit/posts/views.py
@@ -10,7 +10,7 @@ from django.http import JsonResponse
 # Create your views here.
 
 
-@login_required(login_url='login')
+@login_required(login_url='/login/')
 def create_post(request):
     if request.method == 'POST':
         title = request.POST['title']
@@ -32,50 +32,58 @@ def create_post(request):
 def view_post(request, post_id):
     post = get_object_or_404(Post, id=post_id)
 
-    if request.method == 'POST':
-        content = request.POST.get('content')
-        parent_comment_id = request.POST.get('parent_comment_id')
-        parent_comment = None
+    if request.user.is_authenticated:
+        if request.method == 'POST':
+            content = request.POST.get('content')
+            parent_comment_id = request.POST.get('parent_comment_id')
+            parent_comment = None
 
-        if parent_comment_id:
-            parent_comment = get_object_or_404(Comment, id=parent_comment_id)
+            if parent_comment_id:
+                parent_comment = get_object_or_404(
+                    Comment, id=parent_comment_id)
 
-            # to edit or delete comment
-        if 'edit_comment_id' in request.POST:
-            comment_id = request.POST['edit_comment_id']
-            comment = get_object_or_404(Comment, id=comment_id)
-            if request.user == comment.author:
-                form = CommentForm(request.POST, instance=comment)
-                if form.is_valid():
-                    form.save()
+            # Edit or delete comment
+            if 'edit_comment_id' in request.POST:
+                comment_id = request.POST['edit_comment_id']
+                comment = get_object_or_404(Comment, id=comment_id)
+                if request.user == comment.author:
+                    form = CommentForm(request.POST, instance=comment)
+                    if form.is_valid():
+                        form.save()
+                        return redirect('posts:view_post', post_id=post.id)
+            elif 'delete_comment_id' in request.POST:
+                comment_id = request.POST['delete_comment_id']
+                comment = get_object_or_404(Comment, id=comment_id)
+                if request.user == comment.author:
+                    comment.delete()
                     return redirect('posts:view_post', post_id=post.id)
-        elif 'delete_comment_id' in request.POST:
-            comment_id = request.POST['delete_comment_id']
-            comment = get_object_or_404(Comment, id=comment_id)
-            if request.user == comment.author:
-                comment.delete()
-                return redirect('posts:view_post', post_id=post.id)
-                # edit or delete comment section ends here
-        if content:
-            Comment.objects.create(
-                post=post,
-                author=request.user,
-                content=content,
-                parent_comment=parent_comment
-            )
-            messages.success(request, 'Comment added successfully')
-        else:
-            messages.error(request, 'Text field cannot be empty')
 
-        return redirect('posts:view_post', post_id=post_id)
+            if content:
+                Comment.objects.create(
+                    post=post,
+                    author=request.user,
+                    content=content,
+                    parent_comment=parent_comment
+                )
+                messages.success(request, 'Comment added successfully')
+            else:
+                messages.error(request, 'Text field cannot be empty')
 
-    comments = Comment.objects.filter(post=post, parent_comment__isnull=True).select_related(
-        'author').prefetch_related('replies')
+            return redirect('posts:view_post', post_id=post_id)
 
-    return render(request, 'posts/view_post.html', {'post': post, 'comments': comments})
+        # This block is moved outside the POST check to be accessible for GET requests as well
+        comments = Comment.objects.filter(post=post, parent_comment__isnull=True).select_related(
+            'author').prefetch_related('replies')
+        return render(request, 'posts/view_post.html', {'post': post, 'comments': comments})
+
+    # For unauthenticated users
+    else:
+        comments = Comment.objects.filter(post=post, parent_comment__isnull=True).select_related(
+            'author').prefetch_related('replies')
+        return render(request, 'posts/view_post.html', {'post': post})
 
 
-@login_required(login_url='login')
+@login_required(login_url='/login/')
 def edit_post(request, post_id):
     post = get_object_or_404(Post, pk=post_id)
 
@@ -99,7 +107,7 @@ def edit_post(request, post_id):
     return render(request, 'posts/edit_post.html', {'post': post})
 
 
-@login_required(login_url='login')
+@login_required(login_url='/login/')
 def delete_post(request, post_id):
     post = get_object_or_404(Post, pk=post_id)
 
@@ -112,7 +120,7 @@ def delete_post(request, post_id):
     return redirect('subreddits:view_subreddit', subreddit_id=post.subreddit.id)
 
 
-@login_required(login_url='login')
+@login_required(login_url='/login/')
 def upvote_post(request, post_id):
     post = get_object_or_404(Post, pk=post_id)
     post.upvote(request.user)
@@ -120,7 +128,7 @@ def upvote_post(request, post_id):
     return JsonResponse({'vote_count': post.get_vote_count()})
 
 
-@login_required(login_url='login')
+@login_required(login_url='/login/')
 def downvote_post(request, post_id):
     post = get_object_or_404(Post, pk=post_id)
     post.downvote(request.user)
@@ -128,7 +136,7 @@ def downvote_post(request, post_id):
     return JsonResponse({'vote_count': post.get_vote_count()})
 
 
-@login_required(login_url='login')
+@login_required(login_url='/login/')
 def vote_post(request):
     post_id = request.POST.get('post_id')
     action = request.POST.get('action')
